+++ tests/unit/test_feature.py
@@ -8,6 +8,7 @@ from radish.scenariooutline import ScenarioOutline
 from radish.scenarioloop import ScenarioLoop
 from radish.stepmodel import Step
 from radish.main import setup_config
+from radish.tags import Tag
 
 
 class FeatureTestCase(RadishTestCase):
@@ -131,24 +132,16 @@ class FeatureTestCase(RadishTestCase):
         """
             Test feature's has to run functionality
         """
-        f = Feature(1, "Feature", "Some feature", None, None, [Feature.Tag("foo", None), Feature.Tag("bar", None), Feature.Tag("bad_case", None)])
-        f.has_to_run.when.called_with(None, ["foo"], None).should.return_value(True)
-        f.has_to_run.when.called_with(None, ["good_case", "foo"], None).should.return_value(True)
-        f.has_to_run.when.called_with(None, ["good_case", "bar", "bad_case"], None).should.return_value(True)
-        f.has_to_run.when.called_with(None, ["good_case"], None).should.return_value(False)
+        f = Feature(1, "Feature", "Some feature", None, None)
+        f.has_to_run.when.called_with(None).should.return_value(True)
 
-        f.scenarios.append(Mock(absolute_id=1, has_to_run=lambda x,y,z: False))
-        f.scenarios.append(Mock(absolute_id=2, has_to_run=lambda x,y,z: False))
-        f.scenarios.append(Mock(absolute_id=3, has_to_run=lambda x,y,z: False))
+        f.scenarios.append(Mock(absolute_id=1, has_to_run=lambda x: False))
+        f.scenarios.append(Mock(absolute_id=2, has_to_run=lambda x: False))
+        f.scenarios.append(Mock(absolute_id=3, has_to_run=lambda x: False))
 
-        f.has_to_run.when.called_with([1], None, None).should.return_value(True)
-        f.has_to_run.when.called_with([1, 2], None, None).should.return_value(True)
-        f.has_to_run.when.called_with([3], None, None).should.return_value(True)
-        f.has_to_run.when.called_with([1, 4], None, None).should.return_value(True)
-        f.has_to_run.when.called_with([5, 4], None, None).should.return_value(False)
-        f.has_to_run.when.called_with([6], None, None).should.return_value(False)
-
-        f.has_to_run.when.called_with([1], ["good_case"], None).should.return_value(True)
-        f.has_to_run.when.called_with([1, 2], ["foo", "bad_case"], None).should.return_value(True)
-        f.has_to_run.when.called_with([5, 4], ["bad_case"], None).should.return_value(True)
-        f.has_to_run.when.called_with([6], ["good_case"], None).should.return_value(False)
+        f.has_to_run.when.called_with([1]).should.return_value(True)
+        f.has_to_run.when.called_with([1, 2]).should.return_value(True)
+        f.has_to_run.when.called_with([3]).should.return_value(True)
+        f.has_to_run.when.called_with([1, 4]).should.return_value(True)
+        f.has_to_run.when.called_with([5, 4]).should.return_value(False)
+        f.has_to_run.when.called_with([6]).should.return_value(False)